+++ circleparse/__init__.py
@@ -1,0 +1,3 @@
+from circleparse.replay import parse_replay_file, parse_replay
+
+__version__ = "5.0.1"
