+++ circleparse/enums.py
