--- osrparse/__init__.py
@@ -1,3 +1,0 @@
-from .replay import parse_replay_file, parse_replay
-
-__version__ = "5.0.0"
