--- tests/__init__.py
@@ -1,1 +1,0 @@
-from .replay_test import TestStandardReplay
